% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R
\name{somar}
\alias{somar}
\title{Função para somar}
\usage{
somar(a, b)
}
\arguments{
\item{a}{Um numero}

\item{b}{Outro numero}
}
\description{
Esta função tem a finalidade de somar 2 números
}
\examples{
somar(3, 4)

}
