% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Função para subtrair}
\usage{
subtrair(a, b)
}
\arguments{
\item{a}{Um numero}

\item{b}{Outro numero}
}
\description{
Esta função tem a finalidade de subtrair 2 números
}
\examples{
somar(5, 2)

}
